Move FilterChain to own file
